Move trpc code to trpc dir
